+++ src/app/app.module.ts
@@ -11,6 +11,7 @@ import { HomeModule } from './modules/home/home.module';
 import { RasTokenCallbackModule } from './modules/ras-token-callback/ras-token-callback.module';
 import { LaunchModule } from './modules/launch/launch.module';
 import { InitializeSpinnerModule } from './shared/initialize-spinner/initialize-spinner.module';
+import { AnnouncerModule } from './shared/announcer/announcer.module';
 
 const appearance: MatFormFieldDefaultOptions = {
   appearance: 'outline'
@@ -28,6 +29,7 @@ function initializeApp(settingsService: SettingsService): () => Promise<any> {
   declarations: [AppComponent],
   imports: [
     BrowserModule,
+    AnnouncerModule,
     AppRoutingModule,
     HomeModule,
     RasTokenCallbackModule,