+++ src/app/shared/select-records/select-records.service.ts
@@ -211,9 +211,7 @@ export class SelectRecordsService {
           // Name for display is currently the LOINC short name or dbGap description.
           // To find synonyms, we extended the display name search by additionally
           // searching the long and short LOINC names.
-          query.push(
-            `(${dataFields[key]}:(${value}) OR loinc.SHORTNAME:(${value}) OR loinc.LONG_COMMON_NAME:(${value}))`
-          );
+          query.push(`(${dataFields[key]}:(${value}) OR synonyms:(${value}))`);
         } else {
           query.push(`${dataFields[key]}:(${value})`);
         }